# server/openapi.yaml
openapi: 3.1.0
info:
  title: Voxels public read API
  version: '1.0.0'
  description: |
    The public Voxels API. Most of this is unauthenticated GET.

    Write routes for collection create / deploy / mint writeback are documented
    here. Sign-in, admin, moderation, livekit, radio, metrics, guest passes and
    the internal /grid/* routes are deliberately not described.

    Most handlers answer with an envelope: `{"success": true, "<field>": ...}`,
    where `<field>` is plural for a list and singular for one record. A failed
    lookup answers `{"success": false}`, usually with status 400 and sometimes
    with 200, so check the flag rather than the status.
servers:
  - url: https://www.voxels.com
...
paths:
...
  /federation/parcel/{id}:
    get:
      tags: [federation]
      summary: Read the public parcel genesis hash and latest causal revision
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Base hash, causal clock, parent, active parcel permission revision, preserved edit IDs and pending review IDs
        '400':
          description: Invalid or non-public parcel
...
  /api/parcels.json:
    get:
      operationId: listParcels
      tags: [parcels]
      summary: List parcels, or fetch a batch by id
      description: |
        Without `parcel_ids` this lists minted parcels and each row carries
        `parcel_users`, `owner`, `suburb` and `hash`. With `parcel_ids` it runs a
        different query whose rows are narrower (no owner avatar, no suburb, no
        hash) and whose `y2` is the height, not the top of the box. Ids that are
        not integers are dropped rather than rejected.
      parameters:
        - name: parcel_ids
          in: query
          description: Repeat once per parcel id.
          schema:
            type: array
            items:
              type: integer
        - name: limit
          in: query
          description: Only read when `parcel_ids` is absent.
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcels:
                    type: array
                    items:
                      $ref: '#/components/schemas/ParcelSummary'
...
  /api/parcels/search:
    get:
      operationId: searchParcels
      tags: [parcels]
      summary: Search minted, non-common parcels
      description: |
        `q` is a prefix match on parcel name or address. A bare wallet matches
        the owner exactly instead. No total count, a full page means there may
        be another.
      parameters:
        - name: q
          in: query
          description: Empty matches everything.
          schema:
            type: string
        - name: limit
          in: query
          description: Capped at 50, defaults to 50.
          schema:
            type: integer
        - name: page
          in: query
          description: Zero-based, multiplied by `limit` for the offset.
          schema:
            type: integer
        - name: sort
          in: query
          description: Anything else falls back to `id`.
          schema:
            type: string
            enum: [id, name, height, island, distance]
            default: id
        - name: asc
          in: query
          description: The string `true` sorts ascending, otherwise descending.
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcels:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        name: { type: [string, 'null'] }
                        address: { type: [string, 'null'] }
                        island: { type: [string, 'null'] }
                        kind: { type: [string, 'null'] }
                        x1: { type: integer }
                        x2: { type: integer }
                        y1: { type: integer }
                        y2: { type: integer }
                        z1: { type: integer }
                        z2: { type: integer }
                        height: { type: integer }
  /api/parcels/favorites.json:
    get:
      operationId: listFavoritedParcels
      tags: [parcels]
      summary: Parcels somebody has favorited
      description: |
        `q` matches the wallet that favorited the parcel, not the parcel itself.
        Empty `q` matches every wallet, so the default answer is every favorited
        parcel. The rows come from a join against `favorites` with no grouping,
        so a parcel favorited by several wallets appears once per favorite.
      parameters:
        - name: q
          in: query
          description: Favoriting wallet, matched with `like`. Empty matches all.
          schema:
            type: string
        - name: limit
          in: query
          description: No cap. Missing means no limit.
          schema:
            type: integer
        - name: page
          in: query
          description: Zero-based, multiplied by `limit` for the offset.
          schema:
            type: integer
        - name: sort
          in: query
          description: Anything else falls back to `id` descending.
          schema:
            type: string
            enum: [id, name, height, island, suburb, distance]
            default: id
        - name: asc
          in: query
          description: The string `true` flips the order.
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcels:
                    type: array
                    items:
                      $ref: '#/components/schemas/ParcelSummary'
...
  /api/parcels/{id}.vox:
    get:
      operationId: getParcelVox
      tags: [parcels]
      summary: The parcel's build as a MagicaVoxel file
      parameters:
        - $ref: '#/components/parameters/ParcelId'
      responses:
        '200':
          description: ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Failure'
  /api/parcels/{id}.png:
    get:
      operationId: getParcelImage
      tags: [parcels]
      summary: Redirect to the parcel preview renderer
      description: |
        302 to `/renderer/v1/parcel/{id}.png`, which renders on demand then
        redirects to the CDN thumb when UGC storage is configured.
      parameters:
        - $ref: '#/components/parameters/ParcelId'
      responses:
        '302':
          description: Redirect to the png preview
        '404':
          $ref: '#/components/responses/Failure'
...
  /api/parcels/{id}/history.json:
    get:
      operationId: getParcelHistory
      tags: [parcels]
      summary: Saved versions of a parcel, newest first
      parameters:
        - $ref: '#/components/parameters/ParcelId'
        - name: limit
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: asc
          in: query
          description: The string `true` flips the order.
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: integer
        - name: end_date
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  versions:
                    type: array
                    items:
                      type: object
  /api/parcels/{id}/history-count.json:
    get:
      operationId: getParcelHistoryCount
      tags: [parcels]
      summary: How many versions a parcel has
      parameters:
        - $ref: '#/components/parameters/ParcelId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  info:
                    type: object
  /api/parcels/{id}/history/{version}.json:
    get:
      operationId: getParcelVersion
      tags: [parcels]
      summary: One saved version of a parcel
      parameters:
        - $ref: '#/components/parameters/ParcelId'
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  version:
                    type: object
  /api/parcels/{id}/snapshots.json:
    get:
      operationId: getParcelSnapshots
      tags: [parcels]
      summary: Versions the owner marked as snapshots
      parameters:
        - $ref: '#/components/parameters/ParcelId'
        - name: autosave
          in: query
          description: The string `include` folds autosaves in.
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  snapshots:
                    type: array
                    items:
                      type: object
  /api/suburbs/{suburb_id}/popular.json:
    get:
      operationId: getPopularParcelsBySuburb
      tags: [parcels]
      summary: The busiest parcels in a suburb
      parameters:
        - name: suburb_id
          in: path
          required: true
          schema:
            type: integer
        - name: days
          in: query
          description: Window in days, 7 when absent or unparseable.
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcels:
                    type: array
                    items:
                      type: object
...
components:
  parameters:
    ParcelId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Failure:
      description: The lookup did not land. Some handlers send this with status 200.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Failure'
  schemas:
    Failure:
      type: object
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
...
    AvatarRef:
      description: |
        A citizen, or the bare lowercased wallet string when no avatar row
        matches. Anywhere an owner or author appears, expect either shape.
      oneOf:
        - type: string
        - type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: [string, 'null']
            owner:
              type: string
            created_at:
              type: [string, 'null']
    ParcelUser:
      description: |
        A collaborator on a parcel: their avatar row plus `role`. When they have
        no avatar only `owner` (the lowercased wallet) and `role` are present.
      type: object
      required: [owner, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: [string, 'null']
        owner:
          type: string
        created_at:
          type: [string, 'null']
        role:
          type: string
          enum: [owner, contributor, excluded]
...
    ParcelGeometry:
      description: The parcel footprint as GeoJSON, in EPSG:3857, hundredths of the world unit.
      type: object
      properties:
        type:
          type: string
          const: Polygon
        crs:
          type: object
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
    ParcelBounds:
      description: |
        The parcel's box in metres. x and z are world position, y1 is the floor
        and y2 the ceiling. Watch out for `y2` on the list routes, where several
        queries alias `y2 - y1` to `y2` and hand you the height instead.
      type: object
      properties:
        x1:
          type: number
        x2:
          type: number
        y1:
          type: number
        y2:
          type: number
        z1:
          type: number
        z2:
          type: number
        height:
          type: number
...
    ParcelSummary:
      description: A row from the parcel list routes. No `content`.
      allOf:
        - $ref: '#/components/schemas/ParcelBounds'
        - type: object
          properties:
            id:
              type: integer
            name:
              type: [string, 'null']
            label:
              type: [string, 'null']
            address:
              type: [string, 'null']
            island:
              type: string
            suburb:
              type: [string, 'null']
            geometry:
              $ref: '#/components/schemas/ParcelGeometry'
            owner:
              $ref: '#/components/schemas/AvatarRef'
            parcel_users:
              type: [array, 'null']
              items:
                $ref: '#/components/schemas/ParcelUser'
            hash:
              type: [string, 'null']
            lightmap_url:
              type: [string, 'null']
            visible:
              type: boolean
            distance_to_center:
              type: number
            distance_to_ocean:
              type: number
            distance_to_closest_common:
              type: number
...